+++ ayunis-core-backend/src/iam/orgs/SUMMARY.md
@@ -5,6 +5,6 @@ Manages organizations as the top-level multi-tenancy boundary in Ayunis. Support
 
 This module defines the `Org` entity—the fundamental tenant unit containing a name, associated users, and timestamps. Key use cases include `CreateOrgUseCase` (provisions a new organization, typically during registration), `UpdateOrgUseCase` (renames an organization), `DeleteOrgUseCase` (removes an organization and cascades cleanup), `FindOrgByIdUseCase` (single org lookup), `FindAllOrgIdsUseCase` (batch ID retrieval), and `SuperAdminGetAllOrgsUseCase` (paginated listing for platform administration). The HTTP layer includes a `SuperAdminOrgsController` for admin-only endpoints.
 
-Domain events are defined in `application/events/`: `OrgCreatedEvent` (emitted when a new organization is created during registration, includes the `Org` entity and the founding `User`). The event class exposes a static `EVENT_NAME` string for type-safe listener registration.
+Domain events are defined in `application/events/`: `OrgCreatedEvent` (emitted by `CreateOrgUseCase` whenever a new organization is created — during registration or via super-admin — and includes the `Org` entity). The event class exposes a static `EVENT_NAME` string for type-safe listener registration.
 
 It integrates with **users** (each user references an orgId), **subscriptions** and **trials** (org-scoped billing), **invites** (org-scoped invitations), **teams** (org-scoped team grouping), and **authentication** (org creation during registration). The `OrgsRepositoryPort` abstracts persistence with a PostgreSQL implementation.
